+++ src/engine/RSScreen.cpp
@@ -76,8 +76,8 @@ void RSScreen::init(int width, int height, bool fullscreen){
     // Setup Platform/Renderer backends
     ImGui_ImplSDL2_InitForOpenGL(sdlWindow, gl_context);
     ImGui_ImplOpenGL2_Init();
-    int w = this->height * (4.0f/3.0f);
-    glViewport((this->width - w)/2.0,0,w,this->height);			// Reset The Current Viewport
+    int w = (int) ((float)this->height * (4.0f/3.0f));
+    glViewport((int)((float)(this->width - w)/2.0f),0,w,this->height);			// Reset The Current Viewport
 	glClearColor(0.0f, 0.0f, 0.0f, 1.0f);				// Black Background
     glClear(GL_COLOR_BUFFER_BIT| GL_DEPTH_BUFFER_BIT);
     SDL_ShowWindow(sdlWindow);
